requirements.txt
celery==4.3.0
Django==2.2.7
google-api-python-client==1.7.11
PyYAML==5.1.2
redis==3.3.11
